move ActiveColumn into Medication table
